% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DiscGolf.R
\docType{package}
\name{PoliticalCandidates2019}
\alias{PoliticalCandidates2019}
\alias{PoliticalCandidates2019-package}
\title{2019 political candidates running for office}
\format{
A dataframe with 33925 elements
\describe{
 \item{candidate_id}{unique identifier for each candidate}
  \item{state}{state candidate ran in}
  \item{office_level}{level of office candidate ran for}
  \item{office_role}{specific role candidate ran for}
  \item{electoral_district}{district candidate ran in}
  \item{office_name}{specific name of the office the candidate ran for}
  \item{seat_type}{type of seat the candidate ran for}
  \item{number_elected}{number of times the candidate will have been in office if elected}
  \item{candidate_name}{name of candidate}
  \item{candidate_party}{party of candidate}
  \item{party_roll_up}{party of the candidate on the official ballot}
  \item{race}{race of the candidate}
  \item{gender}{gender of the candidate}
  \item{election_result}{whether or not candidate won their election}
}
}
\source{
\url{https://wholeads.us}
}
\usage{
PoliticalCandidates2019
}
\description{
This package contains political candidates running
for office across all levels of government in 2019

A data set containing political candidates
running for different offices across all levels of
government and their outcomes in 2019
}
\keyword{datasets}
